+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 2.3.0 (?)
 Changes in this release:
  - Fix generation of statistics for empty model files.
+ - Add support for multiple implementations in one implement statement (#141)
 
 # v 2.2.0 (2022-05-04)
 Changes in this release: